refactor: create useSignUpPanelLogic
